Module interpreter_experiments::hmm_eval

Step-by-step evaluation strategy.

“Hmm” stands for “Highly modular microprocedures” (I totally didn’t make that up while writing the docstring).

This primarily exists for when you want to execute a program in small steps to limit for how long it should run, or how fast.

let expr =
    F!(
        'app F!(
            'app F!('lam x . F!('lam y . F![y])),
            'to F![420]
        ),
        'to F![69]
    );
let mut state = HmmState::new(expr, Rc::default());

let mut step = 0;
let result = loop {
    step += 1;

    if step >= 1000 {
        break Err(EvalError::Misc("Too many steps!".into()));
    }

    if let Some(r) = state.step() {
        break r;
    }
};

println!("Result: {:?}", result);
println!("Steps: {:?}", &step);

Output:

Result: Ok(Int(69))
Steps: 11

WARNING: this module is currently a mess!

Structs

Builtin

HmmState

Type Definitions

EvalResult

Expr

Ns

Val